Updated to use the new typed APIs instead of classic
