--- src/main/scala/ClusteringConfig.scala
@@ -1,10 +1,0 @@
-package com.example
-
-import com.typesafe.config.ConfigFactory
-
-object ClusteringConfig {
-  private val config = ConfigFactory.load()
-
-  val clusterName = config.getString("clustering.cluster.name")
-
-}
